@@ -6,6 +6,7 @@ use App\Repository\BusinessCategoryRepository;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\Serializer\Annotation\Groups;
 
 /**
  * @ORM\Entity(repositoryClass=BusinessCategoryRepository::class)
@@ -17,31 +18,34 @@ class BusinessCategory
      * @ORM\GeneratedValue()
      * @ORM\Column(type="integer")
      */
-    private $id;
+    private int $id;
 
     /**
      * @ORM\Column(type="string", length=255)
+     * @Groups({"public_city_homepage"})
      */
-    private $name;
+    private string $name;
 
     /**
      * @ORM\Column(type="string", length=255)
+     * @Groups({"public_city_homepage"})
      */
-    private $slug;
+    private string $slug;
 
     /**
      * @ORM\Column(type="string", length=50, nullable=true)
+     * @Groups({"public_city_homepage"})
      */
-    private $icon;
+    private string $icon;
 
     /**
-     * @ORM\OneToMany(targetEntity=Business::class, mappedBy="businessCategory")
+     * @ORM\OneToMany(targetEntity=Business::class, mappedBy="businessCategory", orphanRemoval=true)
      */
-    private $business;
+    private $businesses;
 
     public function __construct()
     {
-        $this->business = new ArrayCollection();
+        $this->businesses = new ArrayCollection();
     }
 
     public function getId(): ?int
@@ -78,7 +82,7 @@ class BusinessCategory
         return $this->icon;
     }
 
-    public function setIcon(?string $icon): self
+    public function setIcon(string $icon): self
     {
         $this->icon = $icon;
 
@@ -88,15 +92,15 @@ class BusinessCategory
     /**
      * @return Collection|Business[]
      */
-    public function getBusiness(): Collection
+    public function getBusinesses(): Collection
     {
-        return $this->business;
+        return $this->businesses;
     }
 
     public function addBusiness(Business $business): self
     {
-        if (!$this->business->contains($business)) {
-            $this->business[] = $business;
+        if (!$this->businesses->contains($business)) {
+            $this->businesses[] = $business;
             $business->setBusinessCategory($this);
         }
 
@@ -105,8 +109,8 @@ class BusinessCategory
 
     public function removeBusiness(Business $business): self
     {
-        if ($this->business->contains($business)) {
-            $this->business->removeElement($business);
+        if ($this->businesses->contains($business)) {
+            $this->businesses->removeElement($business);
             // set the owning side to null (unless already changed)
             if ($business->getBusinessCategory() === $this) {
                 $business->setBusinessCategory(null);
