@@ -18,35 +18,35 @@ class User implements UserInterface
      * @ORM\GeneratedValue()
      * @ORM\Column(type="integer")
      */
-    private $id;
+    private int $id;
 
     /**
      * @ORM\Column(type="string", length=180, unique=true)
      */
-    private $email;
+    private string $email;
 
     /**
      * @ORM\Column(type="json")
      */
-    private $roles = [];
+    private array $roles = [];
 
     /**
      * @var string The hashed password
      * @ORM\Column(type="string")
      */
-    private $password;
+    private string $password;
 
     /**
      * @var \DateTime
      * @ORM\Column(type="datetime")
      */
-    private $registeredAt;
+    private \DateTime $registeredAt;
 
     /**
      * @var \DateTime
      * @ORM\Column(type="datetime", nullable=true)
      */
-    private $changedPasswordAt;
+    private \DateTime $changedPasswordAt;
 
     /**
      * @ORM\OneToMany(targetEntity=Business::class, mappedBy="user")
